Добавляет задание 5.14.Модуляция (module5-task1) (#4)
